requirements.txt
 
numpy
collections
pandas
seaborn
matplotlib
scikit-surprise
sklearn
random
bleach==3.1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
docutils==0.16
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
jeepney==0.4.3
Jinja2==2.11.2
joblib==0.14.1
keyring==21.2.0
MarkupSafe==1.1.1
pkginfo==1.5.0.1
psycopg2-binary
pycparser==2.20
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
readme-renderer==25.0
requests==2.23.0
requests-toolbelt==0.9.1
scikit-learn==0.22.2.post1
scipy==1.4.1
SecretStorage==3.1.2
six==1.14.0
SQLAlchemy==1.3.16
tqdm==4.43.0
twine==3.1.1
urllib3==1.25.9
webencodings==0.5.1
Werkzeug==1.0.1
